Add files via upload
Final Jumping block code. Used previous knowledge of jumping and sprites along with a guide on how to create a jumping block. Most code was easy to write since it was shown in class previously with most problems coming from implementing the scoring code. Even with the help from the guide the scoring did not register for a lot of the platforms and required a lot of tweaking. Tweaking the movement was also a challenge in the project but was ultimately not too difficult.